Refactor: Move each Resource into its own file
Shortens the app.py to 48 lines. Each Resource has it's own file in the resources folder, should make maintaining the app a bit easier.
